EDU-232 Move shared schemas to shared-schemas.js
